[#2881] dead lock workaround  - pre-load CollectJob
