+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/monitor/DefaultAgentStatMonitor.java
@@ -22,16 +22,13 @@ import com.navercorp.pinpoint.common.util.PinpointThreadFactory;
 import com.navercorp.pinpoint.profiler.context.module.AgentId;
 import com.navercorp.pinpoint.profiler.context.module.AgentStartTime;
 import com.navercorp.pinpoint.profiler.context.module.StatDataSender;
-import com.navercorp.pinpoint.profiler.monitor.collector.AgentStatCollector;
 import com.navercorp.pinpoint.profiler.monitor.collector.AgentStatMetricCollector;
 import com.navercorp.pinpoint.profiler.sender.DataSender;
+import com.navercorp.pinpoint.profiler.sender.EmptyDataSender;
 import com.navercorp.pinpoint.thrift.dto.TAgentStat;
-import com.navercorp.pinpoint.thrift.dto.TAgentStatBatch;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
@@ -76,6 +73,20 @@ public class DefaultAgentStatMonitor implements AgentStatMonitor {
         }
         this.collectionIntervalMs = collectionInterval;
         this.collectJob = new CollectJob(dataSender, agentId, agentStartTimestamp, agentStatCollector, numCollectionsPerBatch);
+
+        preLoadClass(agentId, agentStartTimestamp, agentStatCollector);
+    }
+
+    // https://github.com/naver/pinpoint/issues/2881
+    // #2881 AppClassLoader and PinpointUrlClassLoader Circular dependency deadlock
+    // prevent deadlock for JDK6
+    // Single thread execution is more safe than multi thread execution.
+    // eg) executor.scheduleAtFixedRate(collectJob, 0(initialDelay is zero), this.collectionIntervalMs, TimeUnit.MILLISECONDS);
+    private void preLoadClass(String agentId, long agentStartTimestamp, AgentStatMetricCollector<TAgentStat> agentStatCollector) {
+        logger.debug("pre-load class start");
+        CollectJob collectJob = new CollectJob(EmptyDataSender.INSTANCE, agentId, agentStartTimestamp, agentStatCollector, 1);
+        collectJob.run();
+        logger.debug("pre-load class end");
     }
 
     @Override
@@ -93,71 +104,6 @@ public class DefaultAgentStatMonitor implements AgentStatMonitor {
             Thread.currentThread().interrupt();
         }
         logger.info("AgentStat monitor stopped");
-    }
-
-    // NotThreadSafe
-    private static class CollectJob implements Runnable {
-
-        private final Logger logger = LoggerFactory.getLogger(this.getClass());
-
-        private final DataSender dataSender;
-        private final String agentId;
-        private final long agentStartTimestamp;
-        private final AgentStatMetricCollector<TAgentStat> agentStatCollector;
-        private final int numCollectionsPerBatch;
-
-        // Not thread safe. For use with single thread ONLY
-        private int collectCount = 0;
-        private long prevCollectionTimestamp = System.currentTimeMillis();
-        private List<TAgentStat> agentStats;
-
-        private CollectJob(DataSender dataSender,
-                           String agentId, long agentStartTimestamp,
-                           AgentStatMetricCollector<TAgentStat> agentStatCollector,
-                           int numCollectionsPerBatch) {
-            this.dataSender = dataSender;
-            this.agentId = agentId;
-            this.agentStartTimestamp = agentStartTimestamp;
-            this.agentStatCollector = agentStatCollector;
-            this.numCollectionsPerBatch = numCollectionsPerBatch;
-            this.agentStats = new ArrayList<TAgentStat>(numCollectionsPerBatch);
-        }
-
-        @Override
-        public void run() {
-            final long currentCollectionTimestamp = System.currentTimeMillis();
-            final long collectInterval = currentCollectionTimestamp - this.prevCollectionTimestamp;
-            try {
-                final TAgentStat agentStat = agentStatCollector.collect();
-                agentStat.setTimestamp(currentCollectionTimestamp);
-                agentStat.setCollectInterval(collectInterval);
-                this.agentStats.add(agentStat);
-                if (++this.collectCount >= numCollectionsPerBatch) {
-                    sendAgentStats();
-                    this.collectCount = 0;
-                }
-            } catch (Exception ex) {
-                logger.warn("AgentStat collect failed. Caused:{}", ex.getMessage(), ex);
-            } finally {
-                this.prevCollectionTimestamp = currentCollectionTimestamp;
-            }
-        }
-
-        private void sendAgentStats() {
-            // prepare TAgentStat object.
-            // TODO multi thread issue.
-            // If we reuse TAgentStat, there could be concurrency issue because data sender runs in a different thread.
-            final TAgentStatBatch agentStatBatch = new TAgentStatBatch();
-            agentStatBatch.setAgentId(agentId);
-            agentStatBatch.setStartTimestamp(agentStartTimestamp);
-            agentStatBatch.setAgentStats(this.agentStats);
-            // If we reuse agentStats list, there could be concurrency issue because data sender runs in a different
-            // thread.
-            // So create new list.
-            this.agentStats = new ArrayList<TAgentStat>(numCollectionsPerBatch);
-            logger.trace("collect agentStat:{}", agentStatBatch);
-            dataSender.send(agentStatBatch);
-        }
     }
 
 }
